Separate bigtable::DataClient and bigtable::Table.
This is part of the changes for #101.  It moves the classes to
their own files.  More changes are coming to cleanup the
code duplication and fix the thread safety issue.
