+++ bigtable/client/table_bulk_apply_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "bigtable/client/data.h"
+#include "bigtable/client/table.h"
 
 #include <absl/memory/memory.h>
