+++ bigtable/client/testing/mock_data_client.h
@@ -15,7 +15,7 @@
 #ifndef GOOGLE_CLOUD_CPP_BIGTABLE_CLIENT_TESTING_MOCK_DATA_CLIENT_H_
 #define GOOGLE_CLOUD_CPP_BIGTABLE_CLIENT_TESTING_MOCK_DATA_CLIENT_H_
 
-#include "bigtable/client/data.h"
+#include "bigtable/client/table.h"
 
 #include <google/bigtable/v2/bigtable_mock.grpc.pb.h>
